+++ src/Plutarch/Extra/Identity.hs
@@ -18,12 +18,14 @@ import Plutarch (
     pcon,
     phoistAcyclic,
     plam,
+    pmatch,
     unTermCont,
     (#),
  )
 import Plutarch.Bool (PEq, POrd)
 import Plutarch.Builtin (PIsData)
 import Plutarch.Extra.Applicative (PApplicative (ppure), PApply (pliftA2))
+import Plutarch.Extra.Bind (PBind (pbind))
 import Plutarch.Extra.Boring (PBoring (pboring))
 import Plutarch.Extra.Comonad (
     PComonad (pextract),
@@ -125,3 +127,9 @@ instance PApplicative PIdentity where
 -- | @since 1.2.0
 instance (PBoring a) => PBoring (PIdentity a) where
     pboring = ppure # pboring
+
+-- | @since 1.2.1
+instance PBind PIdentity where
+    pbind = phoistAcyclic $
+        plam $ \xs f ->
+            pmatch xs $ \(PIdentity x) -> f # x